achievements/certificate_verification/non_qr_verification/requirements.txt
numpy
pillow
opencv-python
opencv-contrib-python
requests
tqdm
PyPDF2
pymupdf
pdf2image
reportlab
imagehash
torch
torchvision
scikit-image
matplotlib
pytesseract
python-dateutil
regex
scipy
google-generativeai
poppler-utils
